Set display name of worker processes to layer
Using the [process.title](https://nodejs.org/api/process.html#process_process_title)
method, it's possible to change the title of processes as shown in tools
like `ps`, `top`, and `htop`.

One limitation is that the name can't be longer than the existing one
(see linked documentation), so right now the process name is simply the
layer. This may be confusing in that it won't indicate the process
is a nodejs process. However the ability to gague CPU usage
of individual workers is valuable enough.