completed work to only call country polygon lookup as last resort
The concepts in this code are a bit confusing.  Initially this project would call all layers for polygon lookup, but unfortunately the sheer size of the country polygons was causing a 2.5x performance degradation in the consumer data importers.  The flow is now, once all the non-country layers have been PiP'd:

1.  call country layer if it needs to be (no other layers returned anything)
2.  lookup country by id if country is needed (country is needed for completion)