Only use valid simplified geometries
GEOS requires that polygons have 4 or more points. The last point must
also equal the first point, so with that restriction a 3 point polygon
would either not connect or simply be a line segment. However
simplify.js does not know about this restriction and will gladly
simplify geometries to 1, 2 or 3 points.

If the simplified geometry ends up being 1, 2, or 3 points, just used
the unsimplified one.